+++ Runtime/Scripts/Error.cs
@@ -4,18 +4,18 @@ namespace KindMen.Uxios
 {
     public abstract class Error : Exception
     {
-        public readonly Config Request;
+        public readonly Config Config;
         public readonly Response Response;
 
-        protected Error(string message, Config request, Response response) : base(message)
+        protected Error(string message, Config config, Response response) : base(message)
         {
-            Request = request;
+            Config = config;
             Response = response;
         }
 
-        protected Error(string message, Config request, Response response, Exception innerException) : base(message, innerException)
+        protected Error(string message, Config config, Response response, Exception innerException) : base(message, innerException)
         {
-            Request = request;
+            Config = config;
             Response = response;
         }
     }