+++ Runtime/Scripts/Errors/ConnectionError.cs
@@ -4,7 +4,7 @@ namespace KindMen.Uxios.Errors
 {
     public class ConnectionError : Error
     {
-        public ConnectionError(string message, Config request, Response response) : base(message, request, response)
+        public ConnectionError(string message, Config config, Response response) : base(message, config, response)
         {
         }
     }