+++ Runtime/Scripts/Errors/HttpClientError.cs
@@ -4,7 +4,7 @@ namespace KindMen.Uxios.Errors
 {
     public class HttpClientError : ProtocolError
     {
-        public HttpClientError(string message, Config request, Response response, Exception e) : base(message, request, response, e)
+        public HttpClientError(string message, Config config, Response response, Exception e) : base(message, config, response, e)
         {
         }
     }